backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
